protein_app/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.0.10
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
factory-boy==3.3.0
Faker==19.3.0
idna==3.4
inflection==0.5.1
numpy==1.25.2
packaging==23.1
# pandas==2.0.3
pandas==0.24.2
psycopg2==2.9.7
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
